rename: mapping 엔티티의 위치를 변경 #103
- MemberBook 엔티티와 BookClubMember 엔티티를 mapping 패키지로 이동했습니다.
