+++ be/src/main/java/codesquad/bookkbookk/domain/mapping/entity/BookClubMember.java
@@ -1,4 +1,4 @@
-package codesquad.bookkbookk.domain.book.data.entity;
+package codesquad.bookkbookk.domain.mapping.entity;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -9,32 +9,34 @@ import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 
+import codesquad.bookkbookk.domain.bookclub.data.entity.BookClub;
 import codesquad.bookkbookk.domain.member.data.entity.Member;
 
 import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-@Entity(name = "member_book")
+@Entity(name = "book_club_member")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class MemberBook {
+@AllArgsConstructor
+@Getter
+public class BookClubMember {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "member_book_id")
+    @Column(name = "book_club_member_id")
     private Long id;
-
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "book_club_id")
+    private BookClub bookClub;
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id")
     private Member member;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "book_id")
-    private Book book;
-
-    public MemberBook(Member member, Book book) {
+    public BookClubMember(BookClub bookClub, Member member) {
+        this.bookClub = bookClub;
         this.member = member;
-        this.book = book;
     }
 
 }
-
