+++ be/src/main/java/codesquad/bookkbookk/domain/mapping/repository/BookClubMemberRepository.java
@@ -1,8 +1,8 @@
-package codesquad.bookkbookk.domain.bookclub.repository;
+package codesquad.bookkbookk.domain.mapping.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import codesquad.bookkbookk.domain.bookclub.data.entity.BookClubMember;
+import codesquad.bookkbookk.domain.mapping.entity.BookClubMember;
 
 public interface BookClubMemberRepository extends JpaRepository<BookClubMember, Long> {
